Move utils files in a dedicated folder
