+++ test/unit/component/parser/BpmnParser.test.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { expectPoolLaneEdgeFlowNode } from './json/JsonTestUtils';
+import { expectPoolLaneEdgeFlowNode } from '../../helpers/JsonTestUtils';
 import { newBpmnParser } from '../../../../src/component/parser/BpmnParser';
 import { readFileSync } from '../../../helpers/file-helper';
