+++ test/unit/component/parser/json/BpmnJsonParser.group.test.ts
@@ -21,7 +21,7 @@ import {
   parseJsonAndExpectOnlyPoolsAndFlowNodes,
   parseJsonAndExpectOnlyWarnings,
   parsingMessageCollector,
-} from './JsonTestUtils';
+} from '../../../helpers/JsonTestUtils';
 import { verifyShape } from '../../../helpers/bpmn-model-expect';
 import { ShapeBpmnElementKind } from '../../../../../src/model/bpmn/internal';
 import { GroupUnknownCategoryValueWarning, ShapeUnknownBpmnElementWarning } from '../../../../../src/component/parser/json/warnings';
