+++ test/unit/component/parser/json/BpmnJsonParser.messageFlow.test.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { parseJsonAndExpectOnlyEdges } from './JsonTestUtils';
+import { parseJsonAndExpectOnlyEdges } from '../../../helpers/JsonTestUtils';
 import { verifyEdge } from '../../../helpers/bpmn-model-expect';
 import { Waypoint } from '../../../../../src/model/bpmn/internal/edge/edge';
 import { MessageVisibleKind } from '../../../../../src/model/bpmn/internal';
