+++ test/unit/component/parser/json/BpmnJsonParser.sequenceFlow.conditional.test.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { parseJsonAndExpectOnlyEdgesAndFlowNodes } from './JsonTestUtils';
+import { parseJsonAndExpectOnlyEdgesAndFlowNodes } from '../../../helpers/JsonTestUtils';
 import { verifyEdge } from '../../../helpers/bpmn-model-expect';
 import { SequenceFlowKind } from '../../../../../src/model/bpmn/internal';
 import type { TProcess } from '../../../../../src/model/bpmn/json/baseElement/rootElement/rootElement';
