+++ test/unit/component/parser/json/BpmnJsonParser.sequenceFlow.default.test.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { parseJsonAndExpectOnlyEdgesAndFlowNodes } from './JsonTestUtils';
+import { parseJsonAndExpectOnlyEdgesAndFlowNodes } from '../../../helpers/JsonTestUtils';
 import { verifyEdge } from '../../../helpers/bpmn-model-expect';
 import { SequenceFlowKind } from '../../../../../src/model/bpmn/internal';
 import { Waypoint } from '../../../../../src/model/bpmn/internal/edge/edge';
