+++ test/unit/helpers/TestUtils.ts
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ShapeBpmnElementKind, ShapeUtil } from '../../../../../src/model/bpmn/internal';
-import type BpmnModel from '../../../../../src/model/bpmn/internal/BpmnModel';
-import type Shape from '../../../../../src/model/bpmn/internal/shape/Shape';
+import { ShapeBpmnElementKind, ShapeUtil } from '../../../src/model/bpmn/internal';
+import type BpmnModel from '../../../src/model/bpmn/internal/BpmnModel';
+import type Shape from '../../../src/model/bpmn/internal/shape/Shape';
 
 export const shapeBpmnElementKindForLabelTests = Object.values(ShapeBpmnElementKind)
   .filter(kind => !ShapeUtil.isPoolOrLane(kind))
